ref: Reverse the dependency between breakpad-symbols and minidump-processor
Ideally the processor should not care about the concrete SymbolProvider,
and breakpad-symbols should only implement the required traits.
