+++ breakpad-symbols/Cargo.toml
@@ -18,6 +18,9 @@ travis-ci = { repository = "rust-minidump/rust-minidump" }
 fuzz = []
 # Allow retrieval of symbols via HTTP
 http = ["reqwest", "tempfile"]
+# Allows construction of symbol files from native binaries
+dump_syms = ["dep:dump_syms", "http"]
+# Allows retrieval of CABed symbol files from mozilla servers
 mozilla_cab_symbols = ["http", "cab"]
 
 [dependencies]
@@ -28,6 +31,7 @@ debugid = "0.8.0"
 dump_syms = { version = "1.0.1", optional = true }
 tracing = { version = "0.1.34", features = ["log"] }
 minidump-common = { version = "0.11.0", path = "../minidump-common" }
+minidump-processor = { version = "0.11.0", path = "../minidump-processor" }
 nom = "7"
 range-map = "0.1.5"
 reqwest = { version = "0.11.6", default-features = false, features = [
