+++ breakpad-symbols/src/lib.rs
@@ -37,10 +37,6 @@
 //! }
 //! ```
 
-use async_trait::async_trait;
-use debugid::{CodeId, DebugId};
-use tracing::trace;
-
 use std::boxed::Box;
 use std::collections::HashMap;
 use std::fs;
@@ -48,14 +44,23 @@ use std::path::PathBuf;
 use std::sync::Mutex;
 use std::{borrow::Cow, sync::Arc};
 
-pub use minidump_common::{traits::Module, utils::basename};
-pub use sym_file::walker;
+use async_trait::async_trait;
+use debugid::{CodeId, DebugId};
+use minidump_processor::{FileError, FileKind, FrameSymbolizer, FrameWalker};
+use tracing::trace;
 
-pub use crate::sym_file::{CfiRules, SymbolFile};
+pub use minidump_common::{traits::Module, utils::basename};
+
+pub use sym_file::{walker, CfiRules, SymbolFile};
 
 #[cfg(feature = "http")]
 pub mod http;
+mod multi_provider;
 mod sym_file;
+mod symbolizer;
+
+pub use multi_provider::MultiSymbolProvider;
+pub use symbolizer::Symbolizer;
 
 #[cfg(feature = "http")]
 pub use http::*;
@@ -79,17 +84,6 @@ pub fn simple_symbol_supplier(symbol_paths: Vec<PathBuf>) -> impl SymbolSupplier
 /// to a string containing an entire breakpad .sym file, for tests.
 pub fn string_symbol_supplier(modules: HashMap<String, String>) -> impl SymbolSupplier {
     StringSymbolSupplier::new(modules)
-}
-
-/// Statistics on the symbols of a module.
-#[derive(Default, Debug)]
-pub struct SymbolStats {
-    /// If the module's symbols were downloaded, this is the url used.
-    pub symbol_url: Option<String>,
-    /// If the symbols were found and loaded into memory.
-    pub loaded_symbols: bool,
-    /// If we tried to parse the symbols, but failed.
-    pub corrupt_symbols: bool,
 }
 
 /// A `Module` implementation that holds arbitrary data.
@@ -301,29 +295,6 @@ pub enum SymbolError {
     ParseError(&'static str, u64),
 }
 
-#[derive(Clone, Debug, thiserror::Error)]
-pub enum FileError {
-    #[error("file not found")]
-    NotFound,
-}
-
-/// An error produced by fill_symbol.
-#[derive(Debug)]
-pub struct FillSymbolError {
-    // We don't want to yield a full SymbolError for fill_symbol
-    // as this would involve cloning bulky Error strings every time
-    // someone requested symbols for a missing module.
-    //
-    // As it turns out there's currently no reason to care about *why*
-    // fill_symbol, so for now this is just a dummy type until we have
-    // something to put here.
-    //
-    // The only reason fill_symbol *can* produce an Err is so that
-    // the caller can distinguish between "we had symbols, but this address
-    // didn't map to a function name" and "we had no symbols for that module"
-    // (this is used as a heuristic for stack scanning).
-}
-
 impl PartialEq for SymbolError {
     fn eq(&self, other: &SymbolError) -> bool {
         matches!(
@@ -355,6 +326,8 @@ pub trait SymbolSupplier {
         file_kind: FileKind,
     ) -> Result<PathBuf, FileError>;
 }
+
+pub(crate) type CachedOperation<T, E> = Arc<tokio::sync::OnceCell<Result<T, E>>>;
 
 /// An implementation of `SymbolSupplier` that loads Breakpad text-format symbols from local disk
 /// paths.
@@ -456,40 +429,6 @@ impl SymbolSupplier for StringSymbolSupplier {
     }
 }
 
-/// A trait for setting symbol information on something like a stack frame.
-pub trait FrameSymbolizer {
-    /// Get the program counter value for this frame.
-    fn get_instruction(&self) -> u64;
-    /// Set the name, base address, and paramter size of the function in
-    // which this frame is executing.
-    fn set_function(&mut self, name: &str, base: u64, parameter_size: u32);
-    /// Set the source file and (1-based) line number this frame represents.
-    fn set_source_file(&mut self, file: &str, line: u32, base: u64);
-}
-
-pub trait FrameWalker {
-    /// Get the instruction address that we're trying to unwind from.
-    fn get_instruction(&self) -> u64;
-    /// Check whether the callee has a callee of its own.
-    fn has_grand_callee(&self) -> bool;
-    /// Get the number of bytes the callee's callee's parameters take up
-    /// on the stack (or 0 if unknown/invalid). This is needed for
-    /// STACK WIN unwinding.
-    fn get_grand_callee_parameter_size(&self) -> u32;
-    /// Get a register-sized value stored at this address.
-    fn get_register_at_address(&self, address: u64) -> Option<u64>;
-    /// Get the value of a register from the callee's frame.
-    fn get_callee_register(&self, name: &str) -> Option<u64>;
-    /// Set the value of a register for the caller's frame.
-    fn set_caller_register(&mut self, name: &str, val: u64) -> Option<()>;
-    /// Explicitly mark one of the caller's registers as invalid.
-    fn clear_caller_register(&mut self, name: &str);
-    /// Set whatever registers in the caller should be set based on the cfa (e.g. rsp).
-    fn set_cfa(&mut self, val: u64) -> Option<()>;
-    /// Set whatever registers in the caller should be set based on the return address (e.g. rip).
-    fn set_ra(&mut self, val: u64) -> Option<()>;
-}
-
 /// A simple implementation of `FrameSymbolizer` that just holds data.
 #[derive(Debug, Default)]
 pub struct SimpleFrame {
@@ -536,17 +475,6 @@ impl FrameSymbolizer for SimpleFrame {
     }
 }
 
-/// A type of file related to a module that you might want downloaded.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum FileKind {
-    /// A Breakpad symbol (.sym) file
-    BreakpadSym,
-    /// The native binary of a module ("code file") (.exe/.dll/.so/.dylib...)
-    Binary,
-    /// Extra debuginfo for a module ("debug file") (.pdb/...?)
-    ExtraDebugInfo,
-}
-
 // Can't make Module derive Hash, since then it can't be used as a trait
 // object (because the hash method is generic), so this is a hacky workaround.
 /// A key that uniquely identifies a module:
@@ -565,204 +493,6 @@ fn module_key(module: &(dyn Module + Sync)) -> ModuleKey {
         module.debug_file().map(|s| s.to_string()),
         module.debug_identifier().map(|s| s.to_string()),
     )
-}
-
-/// Symbolicate stack frames.
-///
-/// A `Symbolizer` manages loading symbols and looking up symbols in them
-/// including caching so that symbols for a given module are only loaded once.
-///
-/// Call [`Symbolizer::new`][new] to instantiate a `Symbolizer`. A Symbolizer
-/// requires a [`SymbolSupplier`][supplier] to locate symbols. If you have
-/// symbols on disk in the [customary directory layout][breakpad_sym_lookup], a
-/// [`SimpleSymbolSupplier`][simple] will work.
-///
-/// Use [`get_symbol_at_address`][get_symbol] or [`fill_symbol`][fill_symbol] to
-/// do symbol lookup.
-///
-/// [new]: struct.Symbolizer.html#method.new
-/// [supplier]: trait.SymbolSupplier.html
-/// [simple]: struct.SimpleSymbolSupplier.html
-/// [get_symbol]: struct.Symbolizer.html#method.get_symbol_at_address
-/// [fill_symbol]: struct.Symbolizer.html#method.fill_symbol
-
-type CachedOperation<T, E> = Arc<tokio::sync::OnceCell<Result<T, E>>>;
-
-pub struct Symbolizer {
-    /// Symbol supplier for locating symbols.
-    supplier: Box<dyn SymbolSupplier + Send + Sync + 'static>,
-    /// Cache of symbol locating results.
-    // TODO?: use lru-cache: https://crates.io/crates/lru-cache/
-    // note that using an lru-cache would mess up the fact that we currently
-    // use this for statistics collection. Splitting out statistics would be
-    // way messier but not impossible.
-    symbols: Mutex<HashMap<ModuleKey, CachedOperation<SymbolFile, SymbolError>>>,
-}
-
-impl Symbolizer {
-    /// Create a `Symbolizer` that uses `supplier` to locate symbols.
-    pub fn new<T: SymbolSupplier + Send + Sync + 'static>(supplier: T) -> Symbolizer {
-        Symbolizer {
-            supplier: Box::new(supplier),
-            symbols: Mutex::new(HashMap::new()),
-        }
-    }
-
-    /// Helper method for non-minidump-using callers.
-    ///
-    /// Pass `debug_file` and `debug_id` describing a specific module,
-    /// and `address`, a module-relative address, and get back
-    /// a symbol in that module that covers that address, or `None`.
-    ///
-    /// See [the module-level documentation][module] for an example.
-    ///
-    /// [module]: index.html
-    pub async fn get_symbol_at_address(
-        &self,
-        debug_file: &str,
-        debug_id: DebugId,
-        address: u64,
-    ) -> Option<String> {
-        let k = (debug_file, debug_id);
-        let mut frame = SimpleFrame::with_instruction(address);
-        self.fill_symbol(&k, &mut frame).await.ok()?;
-        frame.function
-    }
-
-    /// Fill symbol information in `frame` using the instruction address
-    /// from `frame`, and the module information from `module`. If you're not
-    /// using a minidump module, you can use [`SimpleModule`][simplemodule] and
-    /// [`SimpleFrame`][simpleframe].
-    ///
-    /// An Error indicates that no symbols could be found for the relevant
-    /// module.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # std::env::set_current_dir(env!("CARGO_MANIFEST_DIR"));
-    /// use std::str::FromStr;
-    /// use debugid::DebugId;
-    /// use breakpad_symbols::{SimpleSymbolSupplier,Symbolizer,SimpleFrame,SimpleModule};
-    ///
-    /// #[tokio::main]
-    /// async fn main() {
-    ///     use std::path::PathBuf;
-    ///     let paths = vec!(PathBuf::from("../testdata/symbols/"));
-    ///     let supplier = SimpleSymbolSupplier::new(paths);
-    ///     let symbolizer = Symbolizer::new(supplier);
-    ///     let debug_id = DebugId::from_str("5A9832E5287241C1838ED98914E9B7FF1").unwrap();
-    ///     let m = SimpleModule::new("test_app.pdb", debug_id);
-    ///     let mut f = SimpleFrame::with_instruction(0x1010);
-    ///     let _ = symbolizer.fill_symbol(&m, &mut f).await;
-    ///     assert_eq!(f.function.unwrap(), "vswprintf");
-    ///     assert_eq!(f.source_file.unwrap(),
-    ///         r"c:\program files\microsoft visual studio 8\vc\include\swprintf.inl");
-    ///     assert_eq!(f.source_line.unwrap(), 51);
-    /// }
-    /// ```
-    ///
-    /// [simplemodule]: struct.SimpleModule.html
-    /// [simpleframe]: struct.SimpleFrame.html
-    pub async fn fill_symbol(
-        &self,
-        module: &(dyn Module + Sync),
-        frame: &mut (dyn FrameSymbolizer + Send),
-    ) -> Result<(), FillSymbolError> {
-        let cached_sym = self.get_symbols(module).await;
-        let sym = cached_sym
-            .get()
-            .unwrap()
-            .as_ref()
-            .map_err(|_| FillSymbolError {})?;
-        sym.fill_symbol(module, frame);
-        Ok(())
-    }
-
-    /// Collect various statistics on the symbols.
-    ///
-    /// Keys are the file name of the module (code_file's file name).
-    pub fn stats(&self) -> HashMap<String, SymbolStats> {
-        self.symbols
-            .lock()
-            .unwrap()
-            .iter()
-            .map(|(k, res)| {
-                let res = res.get().expect("Had uninitialized SymbolFile entry?");
-                let mut stats = SymbolStats::default();
-                match res {
-                    Ok(sym) => {
-                        stats.symbol_url = sym.url.clone();
-                        stats.loaded_symbols = true;
-                        stats.corrupt_symbols = false;
-                    }
-                    Err(SymbolError::NotFound) => {
-                        stats.loaded_symbols = false;
-                    }
-                    Err(SymbolError::MissingDebugFileOrId) => {
-                        stats.loaded_symbols = false;
-                    }
-                    Err(SymbolError::LoadError(_)) => {
-                        stats.loaded_symbols = false;
-                    }
-                    Err(SymbolError::ParseError(..)) => {
-                        stats.loaded_symbols = true;
-                        stats.corrupt_symbols = true;
-                    }
-                }
-                (leafname(&k.0).to_string(), stats)
-            })
-            .collect()
-    }
-
-    /// Tries to use CFI to walk the stack frame of the FrameWalker
-    /// using the symbols of the given Module. Output will be written
-    /// using the FrameWalker's `set_caller_*` APIs.
-    pub async fn walk_frame(
-        &self,
-        module: &(dyn Module + Sync),
-        walker: &mut (dyn FrameWalker + Send),
-    ) -> Option<()> {
-        let cached_sym = self.get_symbols(module).await;
-        let sym = cached_sym.get().unwrap().as_ref();
-        if let Ok(sym) = sym {
-            trace!("found symbols for address, searching for cfi entries");
-            sym.walk_frame(module, walker)
-        } else {
-            trace!("couldn't find symbols for address, cannot use cfi");
-            None
-        }
-    }
-
-    /// Gets the fully parsed SymbolFile for a given module (or an Error).
-    ///
-    /// This returns a CachedOperation which is guaranteed to already be resolved (lifetime stuff).
-    async fn get_symbols(
-        &self,
-        module: &(dyn Module + Sync),
-    ) -> CachedOperation<SymbolFile, SymbolError> {
-        // This clones an Arc<Once> that we will use to only do this operation once
-        let k = module_key(module);
-        let symbol_once = self.symbols.lock().unwrap().entry(k).or_default().clone();
-        symbol_once
-            .get_or_init(|| async {
-                trace!("locating symbols for module {}", module.code_file());
-                self.supplier.locate_symbols(module).await
-            })
-            .await;
-        symbol_once
-    }
-
-    /// Gets the path to a file for a given module (or an Error).
-    ///
-    /// This returns a CachedOperation which is guaranteed to already be resolved (lifetime stuff).
-    pub async fn get_file_path(
-        &self,
-        module: &(dyn Module + Sync),
-        file_kind: FileKind,
-    ) -> Result<PathBuf, FileError> {
-        self.supplier.locate_file(module, file_kind).await
-    }
 }
 
 #[test]
