+++ breakpad-symbols/src/multi_provider.rs
@@ -1,0 +1,80 @@
+use std::collections::HashMap;
+use std::path::PathBuf;
+
+use async_trait::async_trait;
+use minidump_common::traits::Module;
+use minidump_processor::{
+    FileError, FileKind, FillSymbolError, FrameSymbolizer, FrameWalker, SymbolProvider, SymbolStats,
+};
+
+#[derive(Default)]
+pub struct MultiSymbolProvider {
+    providers: Vec<Box<dyn SymbolProvider + Send + Sync>>,
+}
+
+impl MultiSymbolProvider {
+    pub fn new() -> MultiSymbolProvider {
+        Default::default()
+    }
+
+    pub fn add(&mut self, provider: Box<dyn SymbolProvider + Send + Sync>) {
+        self.providers.push(provider);
+    }
+}
+
+#[async_trait]
+impl SymbolProvider for MultiSymbolProvider {
+    async fn fill_symbol(
+        &self,
+        module: &(dyn Module + Sync),
+        frame: &mut (dyn FrameSymbolizer + Send),
+    ) -> Result<(), FillSymbolError> {
+        // Return Ok if *any* symbol provider came back with Ok, so that the user can
+        // distinguish between having no symbols at all and just not being able to
+        // symbolize this particular frame.
+        let mut best_result = Err(FillSymbolError {});
+        for p in self.providers.iter() {
+            let new_result = p.fill_symbol(module, frame).await;
+            best_result = best_result.or(new_result);
+        }
+        best_result
+    }
+
+    async fn walk_frame(
+        &self,
+        module: &(dyn Module + Sync),
+        walker: &mut (dyn FrameWalker + Send),
+    ) -> Option<()> {
+        for p in self.providers.iter() {
+            let result = p.walk_frame(module, walker).await;
+            if result.is_some() {
+                return result;
+            }
+        }
+        None
+    }
+
+    async fn get_file_path(
+        &self,
+        module: &(dyn Module + Sync),
+        file_kind: FileKind,
+    ) -> Result<PathBuf, FileError> {
+        // Return Ok if *any* symbol provider came back with Ok
+        let mut best_result = Err(FileError::NotFound);
+        for p in self.providers.iter() {
+            let new_result = p.get_file_path(module, file_kind).await;
+            best_result = best_result.or(new_result);
+        }
+        best_result
+    }
+
+    fn stats(&self) -> HashMap<String, SymbolStats> {
+        let mut result = HashMap::new();
+        for p in self.providers.iter() {
+            // FIXME: do more intelligent merging of the stats
+            // (currently doesn't matter as only one provider reports non-empty stats).
+            result.extend(p.stats());
+        }
+        result
+    }
+}
